requirements.txt
transformers==4.31.0
sentence-transformers==2.2.2
pinecone-client==2.2.2
datasets==2.14.0
accelerate==0.21.0
einops==0.6.1
langchain==0.0.240
xformers==0.0.20
bitsandbytes==0.41.0
zenml==0.54.0
